pyproject.toml
[project]
name = "pythonproject9-1lesson"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = "^3.14"
authors = [
    { name = "Konstantin Kanin", email = "canin.kostya@gmail.com" }
]

dependencies = [
    "pytest>=9.0.2",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


dynamic = ["authors", "description", "readme"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
packages = [
    { include = "src", from = "." }
]


[tool.poetry.group.dev.dependencies]
requests = "^2.32.5"
pre-commit = "^4.5.1"
pytest-cov = "^7.0.0"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
black = "^25.11.0"
isort = "^7.0.0"
mypy = "^1.19.0"


[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
mypy_path = ["src"]
exclude = [
    ".venv",
    "venv",
    ".git",
    "__pycache__",
    "migrations",
    "dist",
    "build"
]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.black]
line-length = 119
target-version = ["py314"]
include = '\.py$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | __pycache__
  | migrations
  | dist
  | build
)/
'''

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3
force_sort_within_sections = true
skip = [
    ".venv",
    "venv",
    ".git",
    "__pycache__",
    "migrations",
    "dist",
    "build"
]
known_first_party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["tests/test_*.py"]
addopts = "-v --tb=short"

